+++ spec/spec_helper.rb
@@ -21,7 +21,8 @@ PACKAGE_EXPECTATIONS = {
     development_tools:%w[build-essential gdb pkg-config python3-dev python3-pip python3-venv shellcheck],
     container_tools:%w[buildah fuse-overlayfs podman skopeo slirp4netns uidmap],
     java_sdk:%w[maven openjdk-17-jdk-headless],
-    php_sdk:%w[composer php-cli php-common php-curl php-dev php-mbstring php-mysql php-sqlite3 php-xml php-zip]
+    php_sdk:%w[composer php-cli php-common php-curl php-dev php-mbstring php-mysql php-sqlite3 php-xml php-zip],
+    dotnet_sdk:%w[dotnet-sdk-10.0]
   },
   RedHat:{
     baseline:%w[ca-certificates curl git jq lsof procps-ng python3 rsync tar tree unzip],
@@ -29,6 +30,7 @@ PACKAGE_EXPECTATIONS = {
     development_tools:%w[gcc gcc-c++ gdb make pkgconf-pkg-config python3-devel python3-pip],
     container_tools:%w[buildah fuse-overlayfs podman shadow-utils skopeo slirp4netns],
     java_sdk:%w[java-17-openjdk-devel maven],
-    php_sdk:%w[php-cli php-common php-devel php-mbstring php-mysqlnd php-opcache php-pdo php-process php-xml]
+    php_sdk:%w[php-cli php-common php-devel php-mbstring php-mysqlnd php-opcache php-pdo php-process php-xml],
+    dotnet_sdk:%w[dotnet-sdk-10.0]
   }
 }.freeze